+++ src/pages/index.ts
@@ -8,7 +8,7 @@ import User from '@/pages/User';
 import BoothDetail from '@/pages/booth/BoothDetail';
 import TimeTable from '@/pages/performance/TimeTable';
 import PerformanceDetail from '@/pages/performance/PerformanceDetail';
-import LostSearch from '@/pages/main/LostSearch';
+import LostSearch from './main/lost/LostSearch';
 import LostPost from './main/lost/LostPost';
 import Notice from '@/pages/main/notice/Notice';
 import NoticeDetail from '@/pages/main/notice/NoticeDetail';
